accessible/tests/mochitest/tree/test_menu.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible XUL menu hierarchy tests">

  <script type="application/javascript" 
          src="chrome://mochikit/content/MochiKit/packed.js" />
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js" />
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/role.js" />

  <script type="application/javascript">
  <![CDATA[
    ////////////////////////////////////////////////////////////////////////////
    // Test

    function doTest()
    {
      if (LINUX) {
        // XXX: bug 527646

        todo(false, "Failure on linux.");
        SimpleTest.finish();
        return;
      }

      var accTree = {
        role: ROLE_PARENT_MENUITEM,
        name: "menu",
        children: [
          {
            role: ROLE_MENUPOPUP,
            children: [
              {
                role: ROLE_MENUITEM
              },
              {
                role: ROLE_MENUITEM
              }
            ]
          }
        ]
      };
      testAccessibleTree("menu", accTree);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=249292"
         title="Ensure accessible children for toolbarbutton types 'menu' and 'menu-button'">
        Mozilla Bug 249292
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <menu id="menu" label="menu">
        <menupopup>
          <menuitem label="menuitem"/>
          <menuitem label="menuitem"/>
        </menupopup>
      </menu>
    </vbox>
  </hbox>

</window>
